# swagger.yml
openapi: 3.1.0
info:
  title: MedQuest CRM API
  version: 1.0.0
paths:
...
  /requests:
    get:
      tags:
      - Requests
      summary: List Requests
      operationId: list_requests_requests_get
      security:
      - HTTPBearer: []
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
          title: Page
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 20
          title: Limit
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
      - name: assigned_doctor_id
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          title: Assigned Doctor Id
      - name: patient_id
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          title: Patient Id
      - name: priority
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
            maximum: 5
            minimum: 1
          - type: 'null'
          title: Priority
      - name: date_from
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Date From
      - name: date_to
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Date To
      - name: search
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Search
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse_PatientRequestResponse_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - Requests
      summary: Create Request
      operationId: create_request_requests_post
      security:
      - HTTPBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatientRequestCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientRequestResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /requests/{request_id}:
    get:
      tags:
      - Requests
      summary: Get Request
      operationId: get_request_requests__request_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: request_id
        in: path
        required: true
        schema:
          type: integer
          title: Request Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientRequestResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
      - Requests
      summary: Update Request
      operationId: update_request_requests__request_id__put
      security:
      - HTTPBearer: []
      parameters:
      - name: request_id
        in: path
        required: true
        schema:
          type: integer
          title: Request Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatientRequestUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientRequestResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - Requests
      summary: Delete Request
      operationId: delete_request_requests__request_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: request_id
        in: path
        required: true
        schema:
          type: integer
          title: Request Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /requests/{request_id}/status:
    patch:
      tags:
      - Requests
      summary: Change Request Status
      operationId: change_request_status_requests__request_id__status_patch
      security:
      - HTTPBearer: []
      parameters:
      - name: request_id
        in: path
        required: true
        schema:
          type: integer
          title: Request Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangeStatusRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientRequestResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /requests/{request_id}/assign:
    patch:
      tags:
      - Requests
      summary: Assign Doctor
      operationId: assign_doctor_requests__request_id__assign_patch
      security:
      - HTTPBearer: []
      parameters:
      - name: request_id
        in: path
        required: true
        schema:
          type: integer
          title: Request Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignDoctorRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientRequestResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    AssignDoctorRequest:
      properties:
        doctor_id:
          type: integer
          title: Doctor Id
      type: object
      required:
      - doctor_id
      title: AssignDoctorRequest
...
    ChangeStatusRequest:
      properties:
        status:
          type: string
          title: Status
      type: object
      required:
      - status
      title: ChangeStatusRequest
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    PaginatedResponse_PatientRequestResponse_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/PatientRequestResponse'
          type: array
          title: Items
        total:
          type: integer
          title: Total
        page:
          type: integer
          title: Page
        limit:
          type: integer
          title: Limit
      type: object
      required:
      - items
      - total
      - page
      - limit
      title: PaginatedResponse[PatientRequestResponse]
...
    PatientRequestCreate:
      properties:
        patient_id:
          type: integer
          title: Patient Id
        title:
          type: string
          title: Title
        description:
          type: string
          title: Description
        priority:
          type: integer
          maximum: 5.0
          minimum: 1.0
          title: Priority
          default: 3
      type: object
      required:
      - patient_id
      - title
      - description
      title: PatientRequestCreate
    PatientRequestResponse:
      properties:
        id:
          type: integer
          title: Id
        patient_id:
          type: integer
          title: Patient Id
        patient_full_name:
          type: string
          title: Patient Full Name
        title:
          type: string
          title: Title
        description:
          type: string
          title: Description
        status:
          type: string
          title: Status
        priority:
          type: integer
          title: Priority
        assigned_doctor_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Assigned Doctor Id
        assigned_doctor_full_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Assigned Doctor Full Name
        created_by_id:
          type: integer
          title: Created By Id
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - id
      - patient_id
      - patient_full_name
      - title
      - description
      - status
      - priority
      - assigned_doctor_id
      - assigned_doctor_full_name
      - created_by_id
      - created_at
      - updated_at
      title: PatientRequestResponse
    PatientRequestUpdate:
      properties:
        title:
          anyOf:
          - type: string
          - type: 'null'
          title: Title
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        status:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
        priority:
          anyOf:
          - type: integer
            maximum: 5.0
            minimum: 1.0
          - type: 'null'
          title: Priority
        assigned_doctor_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Assigned Doctor Id
      type: object
      title: PatientRequestUpdate
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer